Добавлена регистрация,профиль пользователя, логирование
