+++ onlearn-ui/src/main/resources/application.properties
@@ -10,6 +10,7 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.javax.persistence.validation.mode=none
 
 #spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.xml
 #spring.liquibase.enabled=true
