+++ onlearn-ui/src/main/resources/templates/index.html
@@ -65,19 +65,10 @@
 							</ul>
 						</li>
 						<li><a href="/contacts">Контакты</a></li>
-<!--						<li><a href="/onlearn/student">Профиль студента</a></li>-->
-						<li sec:authorize="hasRole('ADMIN')">
-							<!--							<form 									th:action="@{/student/{login} (login = ${userOnline})}">-->
-<!--								<th:block th:if="${userOnline != null}" > Профиль студента </th:block>-->
+						<li sec:authorize="hasRole('STUDENT')">
 								<a  th:href="@{/student/{login} (login = ${userOnline})}">Профиль студента</a>
-<!--								<button>Профиль</button>-->
-<!--						    </form>-->
-<!--							</sec:authorize>-->
 						</li>
-
 						<li>
-
-<!--							<a href="/onlearn/login">Войти</a>-->
 						<li>
 							<form class="form-inline my-2 my-lg-0" form-method="post" th:action="@{/logout}">
 								<button sec:authorize="isAuthenticated()" class="btn btn-outline-danger my-2 my-sm-0 btn-sm" type="submit">Выйти</button>
